feat: Add OpenAI and TogetherAI as for llms
This commit adds OpenAI and TogetherAI as dependencies for the llms package in the setup.py file. The extras_require section now includes these two packages along with the existing ones. This change was made to support the integration of OpenAI and TogetherAI models in the Large Language Model (LLM) functionality.
