+++ jaclang/core/llms/huggingface.py
@@ -1,6 +1,7 @@
 """Huggingface client for MTLLM."""
 
-import re
+from .base import BaseLLM
+from .utils import logger
 
 PROMPT_TEMPLATE = """
 [System Prompt]
@@ -44,14 +45,14 @@ reflects the expected output type, Follow the format below to provide the output
 """  # noqa E501
 
 
-class Huggingface:
+class Huggingface(BaseLLM):
     """Huggingface API client for Large Language Models (LLMs)."""
 
     MTLLM_PROMPT: str = PROMPT_TEMPLATE
     MTLLM_REASON_SUFFIX: str = WITH_REASON_SUFFIX
     MTLLM_WO_REASON_SUFFIX: str = WITHOUT_REASON_SUFFIX
 
-    def __init__(self, **kwargs: dict) -> None:
+    def __init__(self, verbose: bool = False, **kwargs: dict) -> None:
         """Initialize the Huggingface API client."""
         import torch  # type: ignore
         from transformers import AutoModelForCausalLM, AutoTokenizer, pipeline  # type: ignore
@@ -66,36 +67,24 @@ class Huggingface:
         tokenizer = AutoTokenizer.from_pretrained(
             kwargs.get("model_name", "microsoft/Phi-3-mini-128k-instruct")
         )
+        self.verbose = verbose
         self.pipe = pipeline("text-generation", model=model, tokenizer=tokenizer)
         self.temperature = kwargs.get("temperature", 0.7)
         self.max_tokens = kwargs.get("max_new_tokens", 1024)
 
     def __infer__(self, meaning_in: str, **kwargs: dict) -> str:
         """Infer a response from the input meaning."""
+        if self.verbose:
+            logger.info(f"Meaning In\n{meaning_in}")
         messages = [{"role": "user", "content": meaning_in}]
         output = self.pipe(
             messages,
             temperature=kwargs.get("temperature", self.temperature),
             max_length=kwargs.get("max_new_tokens", self.max_tokens),
-            **kwargs
+            **kwargs,
         )
         return output[0]["generated_text"][-1]["content"]
 
     def __call__(self, input_text: str, **kwargs: dict) -> str:
         """Infer a response from the input text."""
         return self.__infer__(input_text, **kwargs)
-
-    def resolve_output(self, meaning_out: str, did_reason: bool) -> dict:
-        """Resolve the output string to return the reasoning and output."""
-        print(meaning_out)
-        reasoning_match = (
-            re.search(r"\[Reasoning\](.*)\[Output\]", meaning_out)
-            if did_reason
-            else None
-        )
-        output_match = re.search(r"\[Output\](.*)", meaning_out)
-        reasoning = reasoning_match.group(1).strip() if reasoning_match else None
-        output = output_match.group(1).strip() if output_match else None
-        if not output and not reasoning:
-            output = meaning_out.strip()
-        return {"reasoning": reasoning, "output": output}
