+++ jaclang/core/llms/ollama.py
@@ -1,6 +1,7 @@
 """Ollama client for MTLLM."""
 
-import re
+from .base import BaseLLM
+from .utils import logger
 
 PROMPT_TEMPLATE = """
 [System Prompt]
@@ -44,18 +45,19 @@ reflects the expected output type, Follow the format below to provide the output
 """  # noqa E501
 
 
-class Ollama:
+class Ollama(BaseLLM):
     """Ollama API client for Large Language Models (LLMs)."""
 
     MTLLM_PROMPT: str = PROMPT_TEMPLATE
     MTLLM_REASON_SUFFIX: str = WITH_REASON_SUFFIX
     MTLLM_WO_REASON_SUFFIX: str = WITHOUT_REASON_SUFFIX
 
-    def __init__(self, **kwargs: dict) -> None:
+    def __init__(self, verbose: bool = False, **kwargs: dict) -> None:
         """Initialize the Ollama API client."""
         import ollama  # type: ignore
 
         self.client = ollama.Client(host=kwargs.get("host", "http://localhost:11434"))
+        self.verbose = verbose
         self.model_name = kwargs.get("model_name", "phi3")
         self.default_model_params = {
             k: v for k, v in kwargs.items() if k not in ["model_name", "host"]
@@ -63,6 +65,8 @@ class Ollama:
 
     def __infer__(self, meaning_in: str, **kwargs: dict) -> str:
         """Infer a response from the input meaning."""
+        if self.verbose:
+            logger.info(f"Meaning In\n{meaning_in}")
         model = str(kwargs.get("model_name", self.model_name))
         if not self.check_model(model):
             self.download_model(model)
@@ -90,17 +94,3 @@ class Ollama:
     def download_model(self, model_name: str) -> None:
         """Download the model."""
         self.client.pull(model_name)
-
-    def resolve_output(self, meaning_out: str, did_reason: bool) -> dict:
-        """Resolve the output string to return the reasoning and output."""
-        reasoning_match = (
-            re.search(r"\[Reasoning\](.*)\[Output\]", meaning_out)
-            if did_reason
-            else None
-        )
-        output_match = re.search(r"\[Output\](.*)", meaning_out)
-        reasoning = reasoning_match.group(1).strip() if reasoning_match else None
-        output = output_match.group(1).strip() if output_match else None
-        if not output and not reasoning:
-            output = meaning_out.strip()
-        return {"reasoning": reasoning, "output": output}
