+++ jaclang/core/llms/togetherai.py
@@ -42,7 +42,7 @@ WITHOUT_REASON_SUFFIX = """Generate and return the output result(s) only, adheri
 """  # noqa E501
 
 
-class Anthropic(BaseLLM):
+class TogetherAI(BaseLLM):
     """Anthropic API client for MTLLM."""
 
     MTLLM_PROMPT: str = PROMPT_TEMPLATE
@@ -51,11 +51,11 @@ class Anthropic(BaseLLM):
 
     def __init__(self, verbose: bool = False, **kwargs: dict) -> None:
         """Initialize the Anthropic API client."""
-        import anthropic  # type: ignore
+        import together  # type: ignore
 
-        self.client = anthropic.Anthropic()
+        self.client = together.Together()
         self.verbose = verbose
-        self.model_name = kwargs.get("model_name", "claude-3-sonnet-20240229")
+        self.model_name = kwargs.get("model_name", "mistralai/Mistral-7B-Instruct-v0.3")
         self.temperature = kwargs.get("temperature", 0.7)
         self.max_tokens = kwargs.get("max_tokens", 1024)
 
@@ -64,13 +64,13 @@ class Anthropic(BaseLLM):
         if self.verbose:
             logger.info(f"Meaning In\n{meaning_in}")
         messages = [{"role": "user", "content": meaning_in}]
-        output = self.client.messages.create(
+        output = self.client.chat.completions.create(
             model=kwargs.get("model_name", self.model_name),
             temperature=kwargs.get("temperature", self.temperature),
             max_tokens=kwargs.get("max_tokens", self.max_tokens),
             messages=messages,
         )
-        return output.content[0].text
+        return output.choices[0].message.content
 
     def __call__(self, input_text: str, **kwargs: dict) -> str:
         """Infer a response from the input text."""
